seperate routes into different files, add endpoints for voters
