+++ db/schema.sql
@@ -1,3 +1,4 @@
+DROP TABLE IF EXISTS voters;
 DROP TABLE IF EXISTS parties;
 DROP TABLE IF EXISTS candidates;
 CREATE TABLE parties(
@@ -13,4 +14,11 @@ CREATE TABLE candidates (
   party_id INTEGER UNSIGNED,
   CONSTRAINT fk_party FOREIGN KEY (party_id) REFERENCES parties(id) ON DELETE
   SET NULL
+);
+CREATE TABLE voters (
+  id INTEGER PRIMARY KEY,
+  first_name VARCHAR(30) NOT NULL,
+  last_name VARCHAR(30) NOT NULL,
+  email VARCHAR(50) NOT NULL,
+  created_at DATETIME DEFAULT CURRENT_TIMESTAMP
 );
